requirements.txt
astroid==2.3.2
autopep8==1.4.4
colorama==0.4.1
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-heroku==0.3.1
djangorestframework==3.10.3
django-rest-knox==4.1.0
djongo==1.2.36
dnspython==1.16.0
gunicorn==20.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.3
pymongo==3.9.0
pytz==2019.3
six==1.12.0
sqlparse==0.2.4
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2